+++ docs/changelog.md
@@ -6,6 +6,7 @@ layout: doc
 # Changelog
 
 ### 17. Nov 2022
+
 - Images - Face Restoration Page
   - Remove duplicate examples
   - Add CodeFormer
@@ -13,21 +14,25 @@ layout: doc
 - Add Changelog Page
 
 ### 20. Nov 2022
+
 - Images - Add Denoise Page
   - Add Directions example
   - Add Stefantiek example
 
 ### 24. Nov 2022
+
 - Images - Add Deblurring Page
   - Add three GoPro examples
 - Add Image Files Links for Deblurring, Denoise, Face Restoration and Favorites Page in details
 
 ### 25. Nov 2022
+
 - Images - Update Deblurring Page
   - Add DeblurGANv2_MobileNet
   - Add DeblurGANv2_InceptionResNet-v2
 
 ### 01. Dec 2022
+
 - Images - Update Multimodel Page with [Set 5](multimodels.md#set-5)
   - Add Set 5 Photo [Buddy example](multimodels.md#buddy-2)
     - Add HAT versions, Real_HAT_GAN_SRx4 was just released a week ago on 24 Nov 2022  
@@ -41,3 +46,15 @@ layout: doc
   - Add Set 5 Photo [Painting example](multimodels.md#painting-1)
   - Add Set 5 Photo [PC Build example](multimodels.md#pc-build-1)
   - Add Set 5 Photo [Snowboard example](multimodels.md#snowboard)
+
+### 15. Dec 2022
+
+- Images - Update Multimodels Page
+  - Restructure Page, current (now updatable) examples on Top
+  - Current examples redone in own Vue Image comparison component with vue select and fore rerender component when selection happens so that multi-image comparison is possible. The list is built by traversing the folders with nodejs and saving filenames and paths in a data.json file so the frontend can use it on page load, the images get bundled with the application. Now I can simply manage files and these examples will keep themselves updated with each new build (drag&dropping 300+ images through the imgsli website to update a single example is now in the past). With the new comparison component it is simpler to drag the image around than to try to hit the middle slider to move it around. With the mouse wheel one can zoom in up to a hardcoded limit.
+  - Added AI Generated Images examples with vue component
+  - Added Anime examples with the vue component. Models from the 'Anime' Section in the Upscale Wiki Model Database have been used additionally.
+- Images - Remove single model pages
+- Images - Models used page is now also automatically generated thorugh traversing files in some example directories since I generally named the output files according to the model used
+- Images - update all examples from imgsli to new local vue component
+- Images - Bundle all image sources files inside the app